添加获取时长，设置分辨率

+++ black_board_sdk/src/bf/cloud/android/components/mediaplayer/MediaControllerBase.java
@@ -6,8 +6,8 @@ public abstract class MediaControllerBase {
         void	stop();
         void    pause();
         void	resume();
-        int     getDuration();
-        int     getCurrentPosition();
+        long     getDuration();
+        long     getCurrentPosition();
         void    seekTo(int pos);
         boolean isPlaying();
         int     getBufferPercentage();
